rafi_deploy/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rafi_deploy"
version = "0.1.0"
description = "Onboarding, config generation, and deploy automation for Rafi AI Assistant"
requires-python = ">=3.11"
dependencies = [
    "deepgram-sdk==3.7.0",
    "openai==1.52.0",
    "twilio==9.3.6",
    "supabase==2.10.0",
    "httpx==0.27.2",
    "pyyaml==6.0.2",
    "pydantic==2.9.2",
    "paramiko==3.5.0",
    "sounddevice==0.5.0",
    "soundfile==0.12.1",
    "python-json-logger==2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
]

[project.scripts]
rafi-deploy = "src.cli:main"

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (requires external API credentials)",
    "security: Security tests (sanitization, injection prevention)",
    "e2e: End-to-end tests (requires full test environment)",
]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
